alternate solution for escape-hatches/need-effect challenge 3
giving <EditContact> a key was my first thought, but not having control over the component that calls it made that approach difficult. after reviewing the solution given, it seems my approach was correct but i didn't think of creating a separate component meant to be between the caller and <EditContact>. added for completeness.